[#55648798] Fix crash on startup of OpenStudio
@elainethale fyi if you are not already aware of this, the issue was
discovered after I merged in develop branch.

+++ openstudiocore/src/project/CloudSettingsRecord.hpp
@@ -66,7 +66,7 @@ OPENSTUDIO_ENUM(CloudSettingsRecordColumns,
   ((serverPath)(TEXT)(10))
   ((serverUrlRecordId)(INTEGER)(11))
   ((workerPath)(TEXT)(12))
-  ((workerUrlRecordId)(13))
+  ((workerUrlRecordId)(INTEGER)(13))
   ((haltOnStop)(BOOLEAN)(14))
   ((username)(TEXT)(15))
 );
